+++ btc-withdrawal/src/main/kotlin/com/d3/btc/withdrawal/config/BtcWithdrawalAppConfiguration.kt
@@ -1,0 +1,156 @@
+package com.d3.btc.withdrawal.config
+
+import com.d3.btc.fee.BtcFeeRateService
+import com.d3.btc.provider.BtcRegisteredAddressesProvider
+import config.BitcoinConfig
+import config.loadConfigs
+import io.grpc.ManagedChannelBuilder
+import jp.co.soramitsu.iroha.java.IrohaAPI
+import jp.co.soramitsu.iroha.java.QueryAPI
+import model.IrohaCredential
+import org.bitcoinj.wallet.Wallet
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import provider.NotaryPeerListProviderImpl
+import sidechain.iroha.IrohaChainListener
+import sidechain.iroha.consumer.IrohaConsumerImpl
+import sidechain.iroha.util.ModelUtil
+import com.d3.btc.withdrawal.provider.BtcChangeAddressProvider
+import com.d3.btc.withdrawal.provider.BtcWhiteListProvider
+import com.d3.btc.withdrawal.statistics.WithdrawalStatistics
+import java.io.File
+import java.util.concurrent.Executors
+
+val withdrawalConfig =
+    loadConfigs("btc-withdrawal", BtcWithdrawalConfig::class.java, "/btc/withdrawal.properties").get()
+
+@Configuration
+class BtcWithdrawalAppConfiguration {
+
+    private val withdrawalKeypair = ModelUtil.loadKeypair(
+        withdrawalConfig.withdrawalCredential.pubkeyPath,
+        withdrawalConfig.withdrawalCredential.privkeyPath
+    ).fold({ keypair -> keypair }, { ex -> throw ex })
+
+    private val signatureCollectorKeypair = ModelUtil.loadKeypair(
+        withdrawalConfig.signatureCollectorCredential.pubkeyPath,
+        withdrawalConfig.signatureCollectorCredential.privkeyPath
+    ).fold({ keypair -> keypair }, { ex -> throw ex })
+
+    private val btcRegistrationCredential = ModelUtil.loadKeypair(
+        withdrawalConfig.registrationCredential.pubkeyPath,
+        withdrawalConfig.registrationCredential.privkeyPath
+    ).fold({ keypair ->
+        IrohaCredential(withdrawalConfig.registrationCredential.accountId, keypair)
+    }, { ex -> throw ex })
+
+    private val btcFeeRateCredential = ModelUtil.loadKeypair(
+        withdrawalConfig.btcFeeRateCredential.pubkeyPath,
+        withdrawalConfig.btcFeeRateCredential.privkeyPath
+    ).fold({ keypair ->
+        IrohaCredential(withdrawalConfig.btcFeeRateCredential.accountId, keypair)
+    }, { ex -> throw ex })
+
+    @Bean
+    fun withdrawalStatistics() = WithdrawalStatistics.create()
+
+    @Bean
+    fun wallet() = Wallet.loadFromFile(File(withdrawalConfig.bitcoin.walletPath))
+
+    @Bean
+    fun withdrawalCredential() =
+        IrohaCredential(withdrawalConfig.withdrawalCredential.accountId, withdrawalKeypair)
+
+    @Bean
+    fun withdrawalConsumer() = IrohaConsumerImpl(withdrawalCredential(), irohaAPI())
+
+    @Bean
+    fun signatureCollectorCredential() =
+        IrohaCredential(withdrawalConfig.signatureCollectorCredential.accountId, signatureCollectorKeypair)
+
+    @Bean
+    fun signatureCollectorConsumer() = IrohaConsumerImpl(signatureCollectorCredential(), irohaAPI())
+
+    @Bean
+    fun withdrawalConfig() = withdrawalConfig
+
+    @Bean
+    fun irohaAPI(): IrohaAPI {
+        val irohaAPI = IrohaAPI(withdrawalConfig.iroha.hostname, withdrawalConfig.iroha.port)
+        /**
+         * It's essential to handle blocks in this service one-by-one.
+         * This is why we explicitly set single threaded executor.
+         */
+        irohaAPI.setChannelForStreamingQueryStub(
+            ManagedChannelBuilder.forAddress(
+                withdrawalConfig.iroha.hostname,
+                withdrawalConfig.iroha.port
+            ).executor(Executors.newSingleThreadExecutor()).usePlaintext().build()
+        )
+        return irohaAPI
+    }
+
+    @Bean
+    fun withdrawalIrohaChainListener() = IrohaChainListener(
+        irohaAPI(),
+        withdrawalCredential()
+    )
+
+    @Bean
+    fun queryAPI() = QueryAPI(irohaAPI(), btcFeeRateCredential.accountId, btcFeeRateCredential.keyPair)
+
+    @Bean
+    fun btcRegisteredAddressesProvider(): BtcRegisteredAddressesProvider {
+        return BtcRegisteredAddressesProvider(
+            QueryAPI(irohaAPI(), btcRegistrationCredential.accountId, btcRegistrationCredential.keyPair),
+            withdrawalConfig.registrationCredential.accountId,
+            withdrawalConfig.notaryCredential.accountId
+        )
+    }
+
+    @Bean
+    fun btcFeeRateAccount() = withdrawalConfig.btcFeeRateCredential.accountId
+
+    @Bean
+    fun btcFeeRateService() =
+        BtcFeeRateService(
+            IrohaConsumerImpl(
+                btcFeeRateCredential, irohaAPI()
+            ),
+            btcFeeRateCredential.accountId, queryAPI()
+        )
+
+    @Bean
+    fun withdrawalQueryAPI() = QueryAPI(irohaAPI(), withdrawalCredential().accountId, withdrawalCredential().keyPair)
+
+    @Bean
+    fun whiteListProvider(): BtcWhiteListProvider {
+        return BtcWhiteListProvider(
+            withdrawalConfig.registrationCredential.accountId,
+            withdrawalQueryAPI()
+        )
+    }
+
+    @Bean
+    fun btcChangeAddressProvider(): BtcChangeAddressProvider {
+        return BtcChangeAddressProvider(
+            withdrawalQueryAPI(),
+            withdrawalConfig.mstRegistrationAccount,
+            withdrawalConfig.changeAddressesStorageAccount
+        )
+    }
+
+    @Bean
+    fun blockStoragePath() = withdrawalConfig().bitcoin.blockStoragePath
+
+    @Bean
+    fun btcHosts() = BitcoinConfig.extractHosts(withdrawalConfig().bitcoin)
+
+    @Bean
+    fun notaryPeerListProvider() =
+        NotaryPeerListProviderImpl(
+            queryAPI(),
+            withdrawalConfig.notaryListStorageAccount,
+            withdrawalConfig.notaryListSetterAccount
+        )
+}
