+++ notary-btc-integration-test/src/integration-test/kotlin/integration/btc/BtcFullPipelineTest.kt
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.fail
 import sidechain.iroha.CLIENT_DOMAIN
 import util.getRandomString
 import util.hex
-import withdrawal.btc.handler.CurrentFeeRate
+import com.d3.btc.withdrawal.handler.CurrentFeeRate
 import java.io.File
 import java.math.BigDecimal
 import java.security.KeyPair
